#3 Reorganize api folder
