+++ src/Router.js
@@ -8,7 +8,7 @@ import { LoginScreen } from "./screens/LoginScreen/LoginScreen";
 import Firebase from "./api/firebase/Firebase";
 import * as ROUTES from "./routes";
 import {GoogleLoginResponse} from "react-google-login";
-import { FirebaseDatabase } from "./api/FirebaseDatabase";
+import { FirebaseDatabase } from "./api/storage/firebase/FirebaseDatabase";
 
 Firebase.initialize();
 
@@ -22,7 +22,7 @@ const Router = () => {
     const authResponseHandler = async (authResponse) => {
         setAuthObject(authResponse);
 
-        const { id_token, access_token } = authResponse?.tokenObj;
+        const { id_token, access_token } = authResponse?.tokenObj ?? {id_token: null, access_token: null};
 
         if (id_token && access_token) {
             const firebaseUserResult = await Firebase.authenticateWithTokens(id_token, access_token);
@@ -33,6 +33,8 @@ const Router = () => {
             //Store the access token and refresh token in the database so that
             //the backend can make Gdrive requests for the user later
             await FirebaseDatabase.pushToUserRef("tokens", {accessToken: access_token, refreshToken: refreshToken});
+
+
 
             if (authResponse) {
                 setUserLoggedIn(true);
