+++ src/api/storage/firebase/QuotesFirebaseStorage.js
@@ -1,0 +1,5 @@
+
+
+export async function getQuotesList() {
+    return [];
+}
